% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\name{add_extra_time_columns}
\alias{add_extra_time_columns}
\title{add_extra_time_columns}
\usage{
add_extra_time_columns(df, timestamp_column = "timestamp")
}
\arguments{
\item{df}{arrow object/dataframe}

\item{timestamp_column}{optional name of the timestamp column, default is "timestamp"}
}
\value{
df/arrow object with added columns jahr, yday, timestamp_hours_only
}
\description{
add_extra_time_columns
}
\examples{
\dontrun{
add_extra_time_columns(miv_data)
add_extra_time_columns(miv_data, "name_of_timestamp_column")
}
}
\keyword{internal}
