% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangle.R
\name{add_tooltip_location}
\alias{add_tooltip_location}
\title{add_tooltip_location}
\usage{
add_tooltip_location(df_location)
}
\arguments{
\item{df_location}{data.frame with locations, needs to have variables loc_name and loc_id}
}
\value{
dataframe with additional tooltip column
}
\description{
Add tooltip to location dataframe
}
\examples{
\dontrun{
add_tooltip_location(df_location)
}
}
\keyword{internal}
