% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_daily.R
\name{aggregate_to_daily}
\alias{aggregate_to_daily}
\title{aggregate_to_daily}
\usage{
aggregate_to_daily(input_data, min_n_per_day)
}
\arguments{
\item{input_data}{velo/miv tibble with hourly data}

\item{min_n_per_day}{int, minimal number of hours to be present for a day to be valid/included}
}
\value{
tibble with one entry per day for each direction
}
\description{
aggregate_to_daily
}
\keyword{internal}
