% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_wrangle.R
\name{categorise_weekdays}
\alias{categorise_weekdays}
\title{categorise_weekdays}
\usage{
categorise_weekdays(df, feiertage, name_date_column = "timestamp")
}
\arguments{
\item{df}{tibble to be categorised}

\item{feiertage}{ogd dataset}

\item{name_date_column}{name of the date/datetime column, default "timestamp"}
}
\value{
df with new column day_category
}
\description{
categorise days into workdays, Saturdays, Sundays/bank holidays
}
\examples{
\dontrun{
categorise_weekdays(df_data, df_feiertage)
}
}
\keyword{internal}
