% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_velo_directions.R
\name{clean_velo_directions}
\alias{clean_velo_directions}
\title{clean_velo_directions}
\usage{
clean_velo_directions(velos)
}
\arguments{
\item{velos}{dataset for velos, tibble}
}
\value{
tibble with cleaned directions for locations where the directions have been switched (in memory)
}
\description{
Function to correct for cases, when bike counters have been replaced and the directions of
velo_in and velo_out have been switched
}
\examples{
\dontrun{
velo_cleaned <- clean_velo_directions(velos)
}
}
\keyword{internal}
