% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_velo_directions.R
\name{clean_velo_directions_one_location}
\alias{clean_velo_directions_one_location}
\title{clean_velo_directions_one_location}
\usage{
clean_velo_directions_one_location(loc_id_selected, velos)
}
\arguments{
\item{loc_id_selected}{string, selected loc_id}

\item{velos}{velo dataset}
}
\value{
in-memory cleaned dataset for the selected loc_id only
}
\description{
helper function to actually switch directions for one location
}
\examples{
\dontrun{
clean_velo_directions_one_location("VZS_BINZ", velos)
}
}
