% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\name{get_current_ids}
\alias{get_current_ids}
\title{get_current_ids}
\usage{
get_current_ids(arrow_data)
}
\arguments{
\item{arrow_data}{arrow dataset/dataframe with "timestamp" and "loc_id" columns}
}
\value{
vector with valid id_locs and date of their latest data
}
\description{
get_current_ids
}
\examples{
\dontrun{
valid_zs <- get_current_ids(velo_all)
}
}
\keyword{internal}
