% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_velo_data.R
\name{pivot_vzs}
\alias{pivot_vzs}
\title{pivot_vzs}
\usage{
pivot_vzs(df, type = c("velo", "fuss"))
}
\arguments{
\item{df}{tibble}

\item{type}{string, velo or fuss}
}
\value{
tibble
}
\description{
function to convert velo/fuss into longer format
}
\keyword{internal}
