% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_jahresentwicklung.R
\name{plot_jahresentwicklung}
\alias{plot_jahresentwicklung}
\title{plot_jahresentwicklung}
\usage{
plot_jahresentwicklung(df, font_name = "", vehicle_type = c("Velos", "Autos"))
}
\arguments{
\item{df}{df with number of vehicles per year (year, n_vehicles) and day_category as well as dir_name and loc_name for labels. If data_id and tooltip are present, a tooltip is added with ggiraph.}

\item{font_name}{optional name of the font to be used in theme; the name used for registering the font for ggplot (with sysfonts::font_add) and/or ggiraph (with systemfonts::register_font)}

\item{vehicle_type}{optional, specifies bike/car to be used in axis label, must be one of "Velos" / "Autos", default "Velos"}
}
\value{
ggplot object
}
\description{
plot yearly development
}
\examples{
\dontrun{
plot_jahresentwicklung(df_jahre)
}
}
