% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locations_stadtkarte.R
\name{plot_locations_stadtkarte}
\alias{plot_locations_stadtkarte}
\title{plot_locations_stadtkarte}
\usage{
plot_locations_stadtkarte(
  df_location,
  stadtkreise,
  duplicates_possible = FALSE,
  raster_color = c("color", "grey")
)
}
\arguments{
\item{df_location}{locations to be plotted as circles, needs loc_id and tooltip columns}

\item{stadtkreise}{sf/tibble with stadtkreise, used for limiting the map and plotting the border}

\item{duplicates_possible}{default FALSE, if TRUE takes the last of several entries for the same location (based on loc_id and column "von")}

\item{raster_color}{default "color", alternative can be "grey" to determine which background map tiles are used}
}
\value{
maplibregl object
}
\description{
plot locations with map tiles in the background
}
\examples{
\dontrun{
plot_locations_stadtkarte(df_loc, read_common_shapes()$kreise, duplicates_possible = TRUE)
}

}
