% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locations_stadtkreise.R
\name{plot_locations_stadtkreise}
\alias{plot_locations_stadtkreise}
\title{plot_locations_stadtkreise}
\usage{
plot_locations_stadtkreise(df, common_shapes, duplicates_possible = FALSE)
}
\arguments{
\item{df}{tibble with geo-information with points to plot (needs column named loc_id)}

\item{common_shapes}{list with "see" and "kreise"}

\item{duplicates_possible}{default FALSE, if TRUE takes the last of several entries for the same location (based on loc_id)}
}
\value{
ggplot object
}
\description{
Function to create an (interactive) ggplot with locations of counting stations, with Stadtkreise as background
}
\examples{
\dontrun{
plot_locations_stadtkreise(df_loc, common_shapes)
}
}
