% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tagesgang.R
\name{plot_tagesgang}
\alias{plot_tagesgang}
\title{plot_tagesgang}
\usage{
plot_tagesgang(df, font_name = "", vehicle_type = c("Velos", "Autos"))
}
\arguments{
\item{df}{data frame with data to be plotted, needs timestamp_hours_only,
n_vehicles, day_category, loc_name and dir_name columns. if data_id and  tooltip columns are present, a tooltip is added with ggiraph}

\item{font_name}{optional name of the font to be used in theme; the name used for registering the font for ggplot (with sysfonts::font_add) and/or ggiraph (with systemfonts::register_font)}

\item{vehicle_type}{optional, specifies bike/car to be used in axis label, must be one of "Velos" / "Autos", default "Velos"}
}
\value{
ggplot object
}
\description{
Function to create a line plot for the "Tagesgang" (number of vehicles over the course of a day) based on the day category (Werktag/Samstag/Sonntag)
}
\examples{
\dontrun{
plot_tagesgang(df_tagesgang)
}
}
