% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_card_data.R
\name{prep_card_data}
\alias{prep_card_data}
\title{Prepare data for bslib cards}
\usage{
prep_card_data(days_over_years, current_year, min_n_per_day)
}
\arguments{
\item{days_over_years}{tibble with miv/velo data, one entry per day (and location and direction; the sum over that day)}

\item{current_year}{int, current year}

\item{min_n_per_day}{int, minimum number of observations per day for the day to be valid}
}
\value{
list of two dataframes, one for sum over latest day, the other for sum over year so far
}
\description{
Function to prepare sum of vehicles on latest valid day and in last year for each location and direction
}
\examples{
\dontrun{
cards_info <- prep_card_data(days_over_years, current_year, min_n_per_day)
}
}
\keyword{internal}
