% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data}
\usage{
prep_data(
  input_data,
  df_location,
  min_n_per_day = 22,
  min_fraction_per_year = 0.75
)
}
\arguments{
\item{input_data}{velo/miv data, tibble}

\item{df_location}{tibble with locations, with loc_id and loc_name}

\item{min_n_per_day}{minimum number of measurements per day to be valid, default 22}

\item{min_fraction_per_year}{minimal fraction of valid days in a year required for a year to be valid, default 0.75}
}
\value{
list of tibbles
}
\description{
prep_data
}
\keyword{internal}
