% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_jahresentwicklung.R
\name{prep_jahresentwicklung}
\alias{prep_jahresentwicklung}
\title{Prepare data for Jahresentwicklung}
\usage{
prep_jahresentwicklung(
  days_over_years,
  min_fraction_per_year,
  feiertage,
  loc_dir
)
}
\arguments{
\item{days_over_years}{miv/velo data tibble with one entry per day (for each location and direction; the sum over that day)}

\item{min_fraction_per_year}{minimal fraction of valid days in a year required for a year to be valid}

\item{feiertage}{tibble with bank holidays based on OGD}

\item{loc_dir}{tibble with all locations and directions}
}
\value{
tibble
}
\description{
Data to create one value per year and day category for each location and direction
}
\examples{
\dontrun{
prep_jahresentwicklung(days_over_years, min_fraction_per_year, feiertage, loc_dir)
}
}
\keyword{internal}
