% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_miv_data.R
\name{prep_miv_data}
\alias{prep_miv_data}
\title{prep_miv_data}
\usage{
prep_miv_data(miv_all)
}
\arguments{
\item{miv_all}{tibble with miv data}
}
\value{
list of tibbles/lists with aggregated data
}
\description{
wrapper around prep_data to provide appropriate location data
}
\examples{
\dontrun{
prep_miv_data(read_miv_data())
}
}
