% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tagesgang.R
\name{prep_tagesgang}
\alias{prep_tagesgang}
\title{prep_tagesgang}
\usage{
prep_tagesgang(input_data, one_year_ago, feiertage)
}
\arguments{
\item{input_data}{miv/velo data, tibble}

\item{one_year_ago}{date/string, one year ago}

\item{feiertage}{tibble with OGD Feiertage}
}
\value{
tibble
}
\description{
prepare data as Tagesgang: one value per hour over 24h
}
\keyword{internal}
