% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_velo_data.R
\name{prep_velo_data}
\alias{prep_velo_data}
\title{prep_velo_data}
\usage{
prep_velo_data(velo_all)
}
\arguments{
\item{velo_all}{tibble with velo data}
}
\value{
list of tibbles/lists with aggregated data
}
\description{
wraps around prep_data, providing the correct velo locations, and runs the additional
cleaning step that is only necessary for bikes but not miv
}
\examples{
\dontrun{
prep_velo_data(read_velo_data())
}
}
