% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_common_shapes.R
\name{read_common_shapes}
\alias{read_common_shapes}
\title{read_common_shapes}
\usage{
read_common_shapes()
}
\value{
named list of shapes for stadtkreise ("kreise"), stadtquartiere ("quartiere"), and the lake
}
\description{
function to read the commonly used shapes
}
\examples{
read_common_shapes()
}
