% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read.R
\name{read_feiertage}
\alias{read_feiertage}
\title{read_feiertage}
\usage{
read_feiertage()
}
\value{
tibble
}
\description{
read Feiertage from OGD and from local file (before 2018, saved as package data)
}
\examples{
df_feier <- read_feiertage()
}
