% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_miv_data.R
\name{read_miv_data}
\alias{read_miv_data}
\title{read_miv_data}
\usage{
read_miv_data()
}
\value{
tibble
}
\description{
function to read miv data; keep only locations that are still active (data within the last year)
}
\examples{
read_miv_data()
}
