% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read_miv.R
\name{read_miv_loc_current}
\alias{read_miv_loc_current}
\title{read currently active MIV locations}
\usage{
read_miv_loc_current()
}
\value{
dataframe with active MIV locations
}
\description{
read currently active MIV locations
}
\examples{
read_miv_loc_current()
}
