% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read_miv.R
\name{read_miv_standorte}
\alias{read_miv_standorte}
\title{read locations of MIV counting}
\usage{
read_miv_standorte()
}
\value{
geo-df with MIV locations
}
\description{
read locations of MIV counting
}
\examples{
read_miv_standorte()
}
