% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_velo_data.R
\name{read_velo_data}
\alias{read_velo_data}
\title{read_velo_data}
\usage{
read_velo_data()
}
\value{
tibble with all velo data
}
\description{
function to read velo data of all years
}
\examples{
read_velo_data()
}
