% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read_velo.R
\name{read_velo_loc_current}
\alias{read_velo_loc_current}
\title{read_velo_loc_current}
\usage{
read_velo_loc_current()
}
\value{
tibble
}
\description{
read locations of velo counts that are currently active
}
\examples{
read_velo_loc_current()
}
