% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_read_velo.R
\name{read_velo_standorte}
\alias{read_velo_standorte}
\title{read_velo_standorte}
\usage{
read_velo_standorte()
}
\value{
sf tibble
}
\description{
read velo/fuss zählungsstandorte from OGD
}
\examples{
\dontrun{
read_velo_standorte()
}
}
