% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszvz-package.R
\docType{package}
\name{sszvz-package}
\alias{sszvz}
\alias{sszvz-package}
\title{sszvz: Package to read data, wrangle, analyse and plot data on Verkehrszählungen from the city of Zurich}
\description{
The package provides functions to read parquet files on the counts of cars and bikes at various points in the city of Zurich; other functions then wrangle, aggregate and plot these data.
}
\author{
\strong{Maintainer}: Sara Hänzi \email{sara.haenzi@zuerich.ch}

Other contributors:
\itemize{
  \item Statistik Stadt Zürich \email{statistik@zuerich.ch} [copyright holder, funder]
}

}
\keyword{internal}
