% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_velo_data.R
\name{sum_and_correct}
\alias{sum_and_correct}
\title{sum_and_correct}
\usage{
sum_and_correct(df)
}
\arguments{
\item{df}{Velo tibble with variables anzahl, fk_standort, id, fk_zaehler, korrekturfaktor, objectid,
loc_id, loc_name, dir_id, dir_name, datum}
}
\value{
tibble with hourly and "corrected" data
}
\description{
sum values from every 15min to hourly values, including applying the correction factor
}
\keyword{internal}
