% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_velo_data.R
\name{sum_to_hourly}
\alias{sum_to_hourly}
\title{sum_to_hourly}
\usage{
sum_to_hourly(df)
}
\arguments{
\item{df}{arrow velo object}
}
\value{
arrow velo but with hourly data
}
\description{
sum the values from every 15min to hourly values; if one value is NA, the hourly value will be NA
}
\details{
round to the nearest hour, see https://lubridate.tidyverse.org/reference/round_date.html
xx.00 = measurements from minutes 0-14
xx.15 = measurements from minutes 15-29
xx.30 = measurements from minutes 30-44
xx.45 = measurements from minutes 45-59
round everything down --> e.g. 8.00 will have measurements from 08:00-08:59
}
\keyword{internal}
